+++ examples/input_processing.rs
@@ -24,7 +24,7 @@ fn spawn_player(mut commands: Commands) {
     let input_map = InputMap::default()
         .with_dual_axis(
             Action::Move,
-            KeyboardVirtualDPad::WASD
+            VirtualDPad::wasd()
                 // You can configure a processing pipeline to handle axis-like user inputs.
                 //
                 // This step adds a circular deadzone that normalizes input values