+++ src/clashing_inputs.rs
@@ -417,7 +417,7 @@ mod tests {
     use bevy::prelude::Reflect;
 
     use super::*;
-    use crate::prelude::{KeyboardVirtualDPad, UserInput};
+    use crate::prelude::{UserInput, VirtualDPad};
     use crate::user_input::ButtonlikeChord;
 
     use crate as leafwing_input_manager;
@@ -456,7 +456,7 @@ mod tests {
             CtrlAltOne,
             ButtonlikeChord::new([ControlLeft, AltLeft, Digit1]),
         );
-        input_map.insert_dual_axis(MoveDPad, KeyboardVirtualDPad::ARROW_KEYS);
+        input_map.insert_dual_axis(MoveDPad, VirtualDPad::arrow_keys());
         input_map.insert(CtrlUp, ButtonlikeChord::new([ControlLeft, ArrowUp]));
 
         input_map
@@ -473,15 +473,14 @@ mod tests {
     }
 
     mod basic_functionality {
+        use super::*;
         use crate::{
             input_map::UpdatedValue,
             plugin::{AccumulatorPlugin, CentralInputStorePlugin},
-            prelude::{AccumulatedMouseMovement, AccumulatedMouseScroll, ModifierKey},
+            prelude::{AccumulatedMouseMovement, AccumulatedMouseScroll, ModifierKey, VirtualDPad},
         };
         use bevy::{input::InputPlugin, prelude::*};
         use Action::*;
-
-        use super::*;
 
         #[test]
         #[ignore = "Figuring out how to handle the length of chords with group inputs is out of scope."]
@@ -501,7 +500,7 @@ mod tests {
             let modified_chord = ButtonlikeChord::modified(ModifierKey::Control, KeyA).decompose();
             assert_eq!(modified_chord.len(), 2);
 
-            let group = KeyboardVirtualDPad::WASD.decompose();
+            let group = VirtualDPad::wasd().decompose();
             assert_eq!(group.len(), 1);
         }
 
@@ -513,11 +512,11 @@ mod tests {
             let ab = ButtonlikeChord::new([KeyA, KeyB]);
             let bc = ButtonlikeChord::new([KeyB, KeyC]);
             let abc = ButtonlikeChord::new([KeyA, KeyB, KeyC]);
-            let axyz_dpad = KeyboardVirtualDPad::new(KeyA, KeyX, KeyY, KeyZ);
-            let abcd_dpad = KeyboardVirtualDPad::WASD;
+            let axyz_dpad = VirtualDPad::new(KeyA, KeyX, KeyY, KeyZ);
+            let abcd_dpad = VirtualDPad::wasd();
 
             let ctrl_up = ButtonlikeChord::new([ArrowUp, ControlLeft]);
-            let directions_dpad = KeyboardVirtualDPad::ARROW_KEYS;
+            let directions_dpad = VirtualDPad::arrow_keys();
 
             assert!(!inputs_clash(a, b));
             assert!(inputs_clash(a, ab.clone()));